Conditionalize (is that a word?) ListProofCheckers on RunMode
