+++ go/libkb/proof_checkers.go
@@ -33,15 +33,5 @@ func NewProofChecker(l RemoteProofChainLink) (ProofChecker, ProofError) {
 	return hook(l)
 }
 
-func ListProofCheckers() (proofCheckers []string) {
-	for k := range _dispatch {
-		// Rooter's a test social network for keybase proofs.
-		if k != "rooter" {
-			proofCheckers = append(proofCheckers, k)
-		}
-	}
-	return proofCheckers
-}
-
 //
 //=============================================================================
